+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Motop 3.1
 ---------
 
 * Avoid crash for operations with missing fields
+* Avoid crash for failed explain, kill actions.
 * Rename column State of the operation block.
 * Show locks with operations.